requirements.txt
Flask==3.1.2
pandas==2.3.1
numpy==1.26.4
scikit-learn==1.6.1
python-dotenv==1.1.1
requests==2.32.4
google-generativeai==0.8.5
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.12
faiss-cpu==1.11.0.post1
Unidecode